HWKAGENT-56 - simplify/eliminate code. we now have single storage adapter. make sure we no longer pass in tenant ID unless in METRIC mode
